requirements.txt
Flask
openai
pillow
frontend
tools 
pdf2image==1.17.0
fitz==0.0.1.dev2
PyMuPDF==1.23.3
boto3